+++ example.js
@@ -91,3 +91,4 @@ var data = {
 const raw = new Buffer(caneri)
 console.log(queclink.parse(raw))
 queclink.parse(raw)
+