+++ src/gmt100.js
@@ -66,8 +66,8 @@ const parse = raw => {
           parsedData[4] !== '' ? parseFloat(parsedData[4]) / 1000 : null,
         ada: parsedData[21] !== '' ? parseFloat(parsedData[21]) / 1000 : null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       odometer: parsedData[19] !== '' ? parseFloat(parsedData[19]) : null
@@ -156,8 +156,8 @@ const parse = raw => {
         inputCharge: null,
         ada: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       odometer: parsedData[19] !== '' ? parseFloat(parsedData[19]) : null
@@ -186,8 +186,8 @@ const parse = raw => {
           parsedData[4] !== '' ? parseFloat(parsedData[4]) / 1000 : null,
         ada: parsedData[21] !== '' ? parseFloat(parsedData[21]) / 1000 : null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       odometer: parsedData[19] !== '' ? parseFloat(parsedData[19]) : null
@@ -216,8 +216,8 @@ const parse = raw => {
           parsedData[4] !== '' ? parseFloat(parsedData[4]) / 1000 : null,
         ada: parsedData[21] !== '' ? parseFloat(parsedData[21]) / 1000 : null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       odometer: parsedData[19] !== '' ? parseFloat(parsedData[19]) : null
@@ -277,8 +277,8 @@ const parse = raw => {
         inputCharge: null,
         ada: null
       },
-      mcc: parsedData[11] !== '' ? parseInt(parsedData[11], 10) : null,
-      mnc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
+      mcc: parsedData[11] !== '' ? utils.latamMcc[parseInt(parsedData[11], 10)] : null,
+      mnc: parsedData[12] !== '' ? utils.getMNC(parsedData[11], parsedData[12]) : null,
       lac: parsedData[13] !== '' ? parseInt(parsedData[13], 16) : null,
       cid: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       odometer: null
@@ -310,8 +310,8 @@ const parse = raw => {
         inputCharge: null,
         ada: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       odometer: null
@@ -338,8 +338,8 @@ const parse = raw => {
         inputCharge: null,
         ada: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       odometer: null
@@ -366,8 +366,8 @@ const parse = raw => {
         inputCharge: null,
         ada: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       odometer: parsedData[18] !== '' ? parseFloat(parsedData[18]) : null
@@ -394,8 +394,8 @@ const parse = raw => {
         inputCharge: null,
         ada: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       odometer: parsedData[18] !== '' ? parseFloat(parsedData[18]) : null
@@ -426,8 +426,8 @@ const parse = raw => {
         inputCharge: null,
         ada: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       odometer: parsedData[18] !== '' ? parseFloat(parsedData[18]) : null
@@ -455,8 +455,8 @@ const parse = raw => {
         inputCharge: null,
         ada: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       odometer: null,