+++ src/gv300w.js
@@ -138,8 +138,8 @@ const parse = raw => {
               ? parseFloat(parsedData[index + 4]) / 1000
               : null
         },
-        mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-        mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+        mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+        mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
         lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
         cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
         satellites:
@@ -328,8 +328,8 @@ const parse = raw => {
             ? parseFloat(parsedData[index + 4]) / 1000
             : null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       satellites:
@@ -653,8 +653,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       odometer: parsedData[19] !== '' ? parseFloat(parsedData[19]) : null,
@@ -685,8 +685,8 @@ const parse = raw => {
         ada: parsedData[21] !== '' ? parseFloat(parsedData[21]) / 1000 : null,
         adb: parsedData[22] !== '' ? parseFloat(parsedData[22]) / 1000 : null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       odometer: parsedData[19] !== '' ? parseFloat(parsedData[19]) : null,
@@ -762,8 +762,8 @@ const parse = raw => {
             ? parseFloat(parsedData[index + 3]) / 1000
             : null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       odometer: parsedData[index] !== '' ? parseFloat(parsedData[index]) : null,
@@ -801,6 +801,32 @@ const parse = raw => {
       odometer: null,
       hourmeter: null
     })
+  } else if (command[1] === 'GTPNR' || command[1] === 'GTPFR') {
+    // Power on/off reason
+    data = Object.assign(data, {
+      alarm: utils.getAlarm(command[1], parsedData[4]),
+      loc: null,
+      speed: null,
+      gpsStatus: null,
+      hdop: null,
+      status: null,
+      azimuth: null,
+      altitude: null,
+      datetime: parsedData[9] !== '' ? utils.parseDate(parsedData[9]) : null,
+      voltage: {
+        battery: null,
+        inputCharge: null,
+        ada: null,
+        adb: null,
+        adc: null
+      },
+      mcc: null,
+      mnc: null,
+      lac: null,
+      cid: null,
+      odometer: null,
+      hourmeter: null
+    })
   } else if (
     command[1] === 'GTMPN' ||
     command[1] === 'GTMPF' ||
@@ -828,8 +854,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[11] !== '' ? parseInt(parsedData[11], 10) : null,
-      mnc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
+      mcc: parsedData[11] !== '' ? utils.latamMcc[parseInt(parsedData[11], 10)] : null,
+      mnc: parsedData[12] !== '' ? utils.getMNC(parsedData[11], parsedData[12]) : null,
       lac: parsedData[13] !== '' ? parseInt(parsedData[13], 16) : null,
       cid: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       odometer: null,
@@ -858,8 +884,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       odometer: null,
@@ -892,8 +918,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       odometer: null,
@@ -922,8 +948,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       odometer: null,
@@ -952,8 +978,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       odometer: parsedData[18] !== '' ? parseFloat(parsedData[18]) : null,
@@ -985,8 +1011,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       odometer: parsedData[18] !== '' ? parseFloat(parsedData[18]) : null,
@@ -1019,8 +1045,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       odometer: parsedData[18] !== '' ? parseFloat(parsedData[18]) : null,
@@ -1050,8 +1076,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       odometer: null,
@@ -1081,8 +1107,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       odometer: null,
@@ -1112,8 +1138,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       odometer: parsedData[20] !== '' ? parseFloat(parsedData[20]) : null,
@@ -1154,8 +1180,8 @@ const parse = raw => {
         ada: null,
         adb: null
       },
-      mcc: parsedData[62] !== '' ? parseInt(parsedData[62], 10) : null,
-      mnc: parsedData[63] !== '' ? parseInt(parsedData[63], 10) : null,
+      mcc: parsedData[62] !== '' ? utils.latamMcc[parseInt(parsedData[62], 10)] : null,
+      mnc: parsedData[63] !== '' ? utils.getMNC(parsedData[62], parsedData[63]) : null,
       lac: parsedData[64] !== '' ? parseInt(parsedData[64], 16) : null,
       cid: parsedData[65] !== '' ? parseInt(parsedData[65], 16) : null,
       odometer: null,
@@ -1272,8 +1298,8 @@ const parse = raw => {
           ada: null,
           adb: null
         },
-        mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-        mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+        mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+        mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
         lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
         cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
         odometer: null,
@@ -1398,8 +1424,8 @@ const parse = raw => {
           ada: null,
           adb: null
         },
-        mcc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
-        mnc: parsedData[17] !== '' ? parseInt(parsedData[17], 10) : null,
+        mcc: parsedData[16] !== '' ? utils.latamMcc[parseInt(parsedData[16], 10)] : null,
+        mnc: parsedData[17] !== '' ? utils.getMNC(parsedData[16], parsedData[17]) : null,
         lac: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
         cid: parsedData[19] !== '' ? parseInt(parsedData[19], 16) : null,
         odometer: null,
@@ -1430,8 +1456,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       odometer: null,