+++ src/gv310lau.js
@@ -141,8 +141,8 @@ const parse = raw => {
               ? parseFloat(parsedData[index + 5]) / 1000
               : null
         },
-        mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-        mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+        mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+        mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
         lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
         cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
         satellites:
@@ -354,8 +354,8 @@ const parse = raw => {
             ? parseFloat(parsedData[index + 5]) / 1000
             : null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       satellites:
@@ -845,8 +845,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites:
@@ -900,8 +900,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites:
@@ -991,8 +991,8 @@ const parse = raw => {
             ? parseFloat(parsedData[index + 5]) / 1000
             : null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       satellites:
@@ -1090,8 +1090,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[11] !== '' ? parseInt(parsedData[11], 10) : null,
-      mnc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
+      mcc: parsedData[11] !== '' ? utils.latamMcc[parseInt(parsedData[11], 10)] : null,
+      mnc: parsedData[12] !== '' ? utils.getMNC(parsedData[11], parsedData[12]) : null,
       lac: parsedData[13] !== '' ? parseInt(parsedData[13], 16) : null,
       cid: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       odometer: null,
@@ -1139,8 +1139,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       satellites:
@@ -1240,8 +1240,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       satellites:
@@ -1284,8 +1284,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       satellites:
@@ -1332,8 +1332,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       satellites:
@@ -1381,8 +1381,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       satellites:
@@ -1427,8 +1427,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       usedSatellites: parsedData[5] !== '' ? parseInt(parsedData[5], 16) : null,
@@ -1474,8 +1474,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       satellites:
@@ -1519,8 +1519,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[63] !== '' ? parseInt(parsedData[63], 10) : null,
-      mnc: parsedData[64] !== '' ? parseInt(parsedData[64], 10) : null,
+      mcc: parsedData[63] !== '' ? utils.latamMcc[parseInt(parsedData[63], 10)] : null,
+      mnc: parsedData[64] !== '' ? utils.getMNC(parsedData[63], parsedData[64]) : null,
       lac: parsedData[65] !== '' ? parseInt(parsedData[65], 16) : null,
       cid: parsedData[66] !== '' ? parseInt(parsedData[66], 16) : null,
       satellites:
@@ -1594,8 +1594,8 @@ const parse = raw => {
           adb: null,
           adc: null
         },
-        mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-        mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+        mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+        mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
         lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
         cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
         satellites:
@@ -1736,8 +1736,8 @@ const parse = raw => {
           adb: null,
           adc: null
         },
-        mcc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
-        mnc: parsedData[17] !== '' ? parseInt(parsedData[17], 10) : null,
+        mcc: parsedData[16] !== '' ? utils.latamMcc[parseInt(parsedData[16], 10)] : null,
+        mnc: parsedData[17] !== '' ? utils.getMNC(parsedData[16], parsedData[17]) : null,
         lac: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
         cid: parsedData[19] !== '' ? parseInt(parsedData[19], 16) : null,
         satellites:
@@ -1781,8 +1781,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       satellites:
@@ -1826,8 +1826,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites:
@@ -1962,8 +1962,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       satellites:
@@ -2417,8 +2417,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites:
@@ -2589,8 +2589,8 @@ const parse = raw => {
       azimuth: parsedData[62] !== '' ? parseFloat(parsedData[62]) : null,
       altitude: parsedData[63] !== '' ? parseFloat(parsedData[63]) : null,
       datetime: parsedData[66] !== '' ? utils.parseDate(parsedData[66]) : null,
-      mcc: parsedData[67] !== '' ? parseInt(parsedData[67], 10) : null,
-      mnc: parsedData[68] !== '' ? parseInt(parsedData[68], 10) : null,
+      mcc: parsedData[67] !== '' ? utils.latamMcc[parseInt(parsedData[67], 10)] : null,
+      mnc: parsedData[68] !== '' ? utils.getMNC(parsedData[67], parsedData[68]) : null,
       lac: parsedData[69] !== '' ? parseInt(parsedData[69], 16) : null,
       cid: parsedData[70] !== '' ? parseInt(parsedData[70], 16) : null,
       satellites:
@@ -2891,8 +2891,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       satellites:
@@ -2941,8 +2941,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       satellites:
@@ -2986,8 +2986,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[20] !== '' ? parseInt(parsedData[20], 10) : null,
-      mnc: parsedData[21] !== '' ? parseInt(parsedData[21], 10) : null,
+      mcc: parsedData[20] !== '' ? utils.latamMcc[parseInt(parsedData[20], 10)] : null,
+      mnc: parsedData[21] !== '' ? utils.getMNC(parsedData[20], parsedData[21]) : null,
       lac: parsedData[22] !== '' ? parseInt(parsedData[22], 16) : null,
       cid: parsedData[23] !== '' ? parseInt(parsedData[23], 16) : null,
       satellites:
@@ -3052,8 +3052,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites:
@@ -3130,8 +3130,8 @@ const parse = raw => {
         adb: null,
         adc: null
       },
-      mcc: parsedData[17] !== '' ? parseInt(parsedData[17], 10) : null,
-      mnc: parsedData[18] !== '' ? parseInt(parsedData[18], 10) : null,
+      mcc: parsedData[17] !== '' ? utils.latamMcc[parseInt(parsedData[17], 10)] : null,
+      mnc: parsedData[18] !== '' ? utils.getMNC(parsedData[17], parsedData[18]) : null,
       lac: parsedData[19] !== '' ? parseInt(parsedData[19], 16) : null,
       cid: parsedData[20] !== '' ? parseInt(parsedData[20], 16) : null,
       satellites: