+++ src/gv58lau.js
@@ -88,18 +88,14 @@ const parse = raw => {
         datetime:
           parsedData[13] !== '' ? utils.parseDate(parsedData[13]) : null,
         voltage: {
-          battery: gnssInfo
-            ? parsedData[index + 7] !== ''
-              ? parseFloat(parsedData[index + 7])
-              : null
-            : parsedData[index + 6] !== ''
-              ? parseFloat(parsedData[index + 6])
+          battery: parsedData[index + 5] !== ''
+              ? parseFloat(parsedData[index + 5])
               : null,
           inputCharge:
             parsedData[4] !== '' ? parseFloat(parsedData[4]) / 1000 : null
         },
-        mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-        mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+        mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+        mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
         lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
         cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
         satellites:
@@ -249,14 +245,14 @@ const parse = raw => {
       datetime: parsedData[14] !== '' ? utils.parseDate(parsedData[14]) : null,
       voltage: {
         battery:
-          parsedData[index + 6] !== ''
-            ? parseFloat(parsedData[index + 6])
+          parsedData[index + 5] !== ''
+            ? parseFloat(parsedData[index + 5])
             : null, // percentage
         inputCharge:
           parsedData[5] !== '' ? parseFloat(parsedData[5]) / 1000 : null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       satellites:
@@ -451,13 +447,9 @@ const parse = raw => {
                   : null
             },
             relay: {
-              configResult:
+              state:
                 parsedData[relIx] !== '' && appendMask[1] === '1'
                   ? parseInt(parsedData[relIx])
-                  : null,
-              state:
-                parsedData[relIx + 1] !== '' && appendMask[1] === '1'
-                  ? parseInt(parsedData[relIx + 1])
                   : null
             }
           }
@@ -653,8 +645,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites:
@@ -739,8 +731,8 @@ const parse = raw => {
         battery: parsedData[4] !== '' ? parseFloat(parsedData[4]) / 1000 : null,
         inputCharge: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites:
@@ -819,8 +811,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[11] !== '' ? parseInt(parsedData[11], 10) : null,
-      mnc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
+      mcc: parsedData[11] !== '' ? utils.latamMcc[parseInt(parsedData[11], 10)] : null,
+      mnc: parsedData[12] !== '' ? utils.getMNC(parsedData[11], parsedData[12]) : null,
       lac: parsedData[13] !== '' ? parseInt(parsedData[13], 16) : null,
       cid: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       satellites:
@@ -905,8 +897,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       satellites:
@@ -1038,8 +1030,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       satellites:
@@ -1117,8 +1109,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       satellites:
@@ -1208,8 +1200,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       satellites:
@@ -1297,8 +1289,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       satellites:
@@ -1383,8 +1375,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       usedSatellites: parsedData[5] !== '' ? parseInt(parsedData[5], 16) : null,
@@ -1464,8 +1456,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       satellites:
@@ -1543,8 +1535,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[63] !== '' ? parseInt(parsedData[63], 10) : null,
-      mnc: parsedData[64] !== '' ? parseInt(parsedData[64], 10) : null,
+      mcc: parsedData[63] !== '' ? utils.latamMcc[parseInt(parsedData[63], 10)] : null,
+      mnc: parsedData[64] !== '' ? utils.getMNC(parsedData[63], parsedData[64]) : null,
       lac: parsedData[65] !== '' ? parseInt(parsedData[65], 16) : null,
       cid: parsedData[66] !== '' ? parseInt(parsedData[66], 16) : null,
       satellites:
@@ -1789,8 +1781,8 @@ const parse = raw => {
           adb: null,
           adc: null
         },
-        mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-        mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+        mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+        mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
         lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
         cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
         satellites:
@@ -1836,8 +1828,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
-      mnc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
+      mcc: parsedData[13] !== '' ? utils.latamMcc[parseInt(parsedData[13], 10)] : null,
+      mnc: parsedData[14] !== '' ? utils.getMNC(parsedData[13], parsedData[14]) : null,
       lac: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       cid: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       satellites:
@@ -1916,8 +1908,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites:
@@ -2004,8 +1996,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[12] !== '' ? parseInt(parsedData[12], 10) : null,
-      mnc: parsedData[13] !== '' ? parseInt(parsedData[13], 10) : null,
+      mcc: parsedData[12] !== '' ? utils.latamMcc[parseInt(parsedData[12], 10)] : null,
+      mnc: parsedData[13] !== '' ? utils.getMNC(parsedData[12], parsedData[13]) : null,
       lac: parsedData[14] !== '' ? parseInt(parsedData[14], 16) : null,
       cid: parsedData[15] !== '' ? parseInt(parsedData[15], 16) : null,
       satellites: includeStatus
@@ -2529,8 +2521,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites:
@@ -2749,8 +2741,8 @@ const parse = raw => {
       azimuth: parsedData[62] !== '' ? parseFloat(parsedData[62]) : null,
       altitude: parsedData[63] !== '' ? parseFloat(parsedData[63]) : null,
       datetime: parsedData[66] !== '' ? utils.parseDate(parsedData[66]) : null,
-      mcc: parsedData[67] !== '' ? parseInt(parsedData[67], 10) : null,
-      mnc: parsedData[68] !== '' ? parseInt(parsedData[68], 10) : null,
+      mcc: parsedData[67] !== '' ? utils.latamMcc[parseInt(parsedData[67], 10)] : null,
+      mnc: parsedData[68] !== '' ? utils.getMNC(parsedData[67], parsedData[68]) : null,
       lac: parsedData[69] !== '' ? parseInt(parsedData[69], 16) : null,
       cid: parsedData[70] !== '' ? parseInt(parsedData[70], 16) : null,
       satellites:
@@ -3085,8 +3077,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
-      mnc: parsedData[16] !== '' ? parseInt(parsedData[16], 10) : null,
+      mcc: parsedData[15] !== '' ? utils.latamMcc[parseInt(parsedData[15], 10)] : null,
+      mnc: parsedData[16] !== '' ? utils.getMNC(parsedData[15], parsedData[16]) : null,
       lac: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       cid: parsedData[18] !== '' ? parseInt(parsedData[18], 16) : null,
       satellites:
@@ -3134,8 +3126,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[20] !== '' ? parseInt(parsedData[20], 10) : null,
-      mnc: parsedData[21] !== '' ? parseInt(parsedData[21], 10) : null,
+      mcc: parsedData[20] !== '' ? utils.latamMcc[parseInt(parsedData[20], 10)] : null,
+      mnc: parsedData[21] !== '' ? utils.getMNC(parsedData[20], parsedData[21]) : null,
       lac: parsedData[22] !== '' ? parseInt(parsedData[22], 16) : null,
       cid: parsedData[23] !== '' ? parseInt(parsedData[23], 16) : null,
       satellites:
@@ -3240,8 +3232,8 @@ const parse = raw => {
         battery: null,
         inputCharge: null
       },
-      mcc: parsedData[14] !== '' ? parseInt(parsedData[14], 10) : null,
-      mnc: parsedData[15] !== '' ? parseInt(parsedData[15], 10) : null,
+      mcc: parsedData[14] !== '' ? utils.latamMcc[parseInt(parsedData[14], 10)] : null,
+      mnc: parsedData[15] !== '' ? utils.getMNC(parsedData[14], parsedData[15]) : null,
       lac: parsedData[16] !== '' ? parseInt(parsedData[16], 16) : null,
       cid: parsedData[17] !== '' ? parseInt(parsedData[17], 16) : null,
       satellites: