+++ src/utils.js
@@ -44,7 +44,8 @@ const devices = {
   '41': 'GV75W',
   FC: 'GV600W',
   '6E': 'GV310LAU',
-  '802004': 'GV58LAU'
+  '802004': 'GV58LAU',
+  '802006': 'GV57CG'
 }
 
 /*
@@ -281,6 +282,21 @@ const gnssTriggerTypes = {
 }
 
 /*
+  MCC List
+*/
+const latamMcc = {
+  716: 'Perú',
+  722: 'Argentina',
+  724: 'Brasil',
+  730: 'Chile',
+  732: 'Colombia',
+  736: 'Bolivia',
+  740: 'Ecuador',
+  744: 'Paraguay',
+  748: 'Uruguay'
+}
+
+/*
   Gets the Queclink Device Type
 */
 const getDevice = raw => {
@@ -296,7 +312,7 @@ const getDevice = raw => {
 const getProtocolVersion = protocol => {
   let deviceType
   let deviceVersion
-  if (protocol.substring(0, 6) === '802004') {
+  if (['802004', '802006'].includes(protocol.substring(0, 6))) {
     deviceType = devices.hasOwnProperty(protocol.substring(0, 6))
       ? devices[protocol.substring(0, 6)]
       : null
@@ -1017,6 +1033,8 @@ const getAlarm = (command, report, extra = false) => {
     return { type: 'Charge', status: false, message: messages[command] }
   } else if (command === 'GTBTC') {
     return { type: 'Charging', status: true, message: messages[command] }
+  } else if (command === 'GTDRM') {
+    return { type: 'Device_Removal', message: messages[command] }
   } else if (command === 'GTSTC') {
     return { type: 'Charging', status: false, message: messages[command] }
   } else if (command === 'GTBPL') {
@@ -1210,22 +1228,10 @@ const getAlarm = (command, report, extra = false) => {
         4: acceleration turning
         5: unknown harsh behavior
       */
-    let reportID = parseInt(report[0], 16)
-    let vel
-    if (reportID === 1) {
-      vel = 'baja'
-    } else if (reportID === 2) {
-      vel = 'media'
-    } else if (reportID === 3) {
-      vel = 'alta'
-    } else {
-      vel = ''
-    }
     const reportType = report[1]
     return {
       type: 'Harsh_Behavior',
       status: parseInt(reportType, 10),
-      velocity: vel !== '' ? vel : null,
       message: messages[command][reportType]
     }
   } else if (command === 'GTHBE') {
@@ -1489,6 +1495,7 @@ module.exports = {
   dTimeStates: dTimeStates,
   dWorkingStates: dWorkingStates,
   gnssTriggerTypes: gnssTriggerTypes,
+  latamMcc: latamMcc,
   getDevice: getDevice,
   getProtocolVersion: getProtocolVersion,
   checkGps: checkGps,
@@ -1512,5 +1519,6 @@ module.exports = {
   hex2dec: hex2dec,
   nHexDigit: nHexDigit,
   sumOnes: sumOnes,
-  parseDate: parseDate
+  parseDate: parseDate,
+  getMNC: getMNC
 }